Update executor
Extract model, update to java-tools, use p2rank to
extrach chains #70.

+++ executor/run_p2rank_task.py
@@ -4,19 +4,15 @@
 #
 import os
 import argparse
-import typing
+import sys
 import json
-import enum
 import datetime
 import subprocess
-import shutil
 import logging
+import shutil
 
-from p2rank_executor import \
-    ExecutorConfiguration, \
-    ConservationType, \
-    OutputType, \
-    execute
+from model import *
+from executor import execute
 
 
 class Status(enum.Enum):
@@ -30,10 +26,16 @@ logging.getLogger().setLevel(logging.DEBUG)
 logger = logging.getLogger("prankweb")
 
 
-def _read_arguments() -> typing.Dict[str, str]:
+def _read_arguments() -> typing.Dict[str, any]:
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "--directory", required=True, help="Directory with task to execute.")
+        "--directory",
+        required=True,
+        help="Directory with task to execute.")
+    parser.add_argument(
+        "--keep-working",
+        action="store_true",
+        help="Keep working data.")
     return vars(parser.parse_args())
 
 
@@ -42,16 +44,16 @@ def main():
     directory = arguments["directory"]
     if not os.path.exists(directory) or not os.path.isdir(directory):
         return
-    execute_directory_task(directory)
+    execute_directory_task(directory, arguments["keep_working"])
 
 
-def execute_directory_task(directory: str):
+def execute_directory_task(directory: str, keep_working: bool):
     log_file = os.path.join(directory, "log")
     with open(log_file, "w", encoding="utf-8") as stream:
-        handler = _create_log_handler(stream)
+        handler = _create_log_handler(sys.stderr)
         logger.addHandler(handler)
         try:
-            _execute_directory_task(directory, stream)
+            _execute_directory_task(directory, sys.stdout, keep_working)
         finally:
             handler.flush()
             logger.removeHandler(handler)
@@ -68,48 +70,45 @@ def _create_log_handler(stream: typing.TextIO):
     return handler
 
 
-def _execute_directory_task(directory: str, stream):
+def _execute_directory_task(directory: str, stream, keep_working: bool):
     status_file = os.path.join(directory, "info.json")
-    status = load_json(status_file)
+    status = _load_json(status_file)
 
     # We can check here if the task is running but, we assume that
     # should we run given task it is regardless of the initial state.
 
     status["status"] = Status.RUNNING.value
-    save_status_file(status_file, status)
+    _save_status_file(status_file, status)
 
-    configuration_file = os.path.join(directory, "input", "configuration.json")
-    user_configuration = load_json(configuration_file)
-
-    structure = user_configuration["structure"]
-    conservation = user_configuration["conservation"]
-
-    pdb_path = os.path.join(directory, "input", structure["file"]) \
-        if "file" in structure and structure["file"] is not None else None
-
-    conservation_type = ConservationType.HMM \
-        if conservation["compute"] else ConservationType.NONE
+    configuration = _load_json(
+        os.path.join(directory, "input", "configuration.json"))
 
     this_directory = os.path.dirname(os.path.realpath(__file__))
 
-     # PROTEIN_UTILS = os.environ.get("PROTEIN_UTILS_CMD", None)
-
-    configuration = ExecutorConfiguration(
-        structure["code"],
-        pdb_path,
-        conservation_type,
-        os.path.join(this_directory, "p2rank.sh"),
-        os.path.join(directory, "working"),
-        os.path.join(directory, "public"),
-        OutputType.PRANKWEB,
-        structure["chains"],
-        structure["sealed"],
-        stream,
-        stream
+    execution = Execution(
+        p2rank=os.path.join(this_directory, "p2rank.sh"),
+        java_tools=os.environ.get("JAVA_TOOLS_CMD", None),
+        working_directory=os.path.join(directory, "working"),
+        output_directory=os.path.join(directory, "public"),
+        output_type=OutputType.PRANKWEB,
+        stdout=stream,
+        stderr=stream,
+        p2rank_configuration=configuration.get("p2rank_configuration", None),
+        structure_code=configuration.get("structure_code", None),
+        structure_file=_structure_path(directory, configuration),
+        structure_uniprot=configuration.get("structure_uniprot", None),
+        structure_sealed=configuration.get("structure_sealed", False),
+        chains=configuration.get("chains", []),
+        conservation=_conservation_type(configuration),
     )
     try:
-        execute(configuration)
+        result = execute(execution)
         status["status"] = Status.SUCCESSFUL.value
+        status["metadata"] = {
+            **status.get("metadata", {}),
+            "predictionName": _output_name(execution),
+            "structureName": result.output_structure_file,
+        }
     except subprocess.CalledProcessError:
         status["status"] = Status.FAILED.value
         logger.exception("External process failed.")
@@ -117,26 +116,58 @@ def _execute_directory_task(directory: str, stream):
         status["status"] = Status.FAILED.value
         logger.exception("Execution failed.")
 
-    shutil.rmtree(os.path.join(directory, "working"))
-    save_status_file(status_file, status)
+    _save_status_file(status_file, status)
+
+    if not keep_working:
+        shutil.rmtree(os.path.join(directory, "working"))
 
 
-def load_json(path: str):
+def _load_json(path: str):
     with open(path, encoding="utf-8") as stream:
         return json.load(stream)
 
 
-def save_status_file(path: str, status: any):
+def _save_status_file(path: str, status: any):
     now = datetime.datetime.today()
     status["lastChange"] = now.strftime('%Y-%m-%dT%H:%M:%S')
-    save_json(path, status)
+    _save_json(path, status)
 
 
-def save_json(path: str, content: any):
+def _save_json(path: str, content: any):
     path_swp = path + ".swp"
     with open(path_swp, "w", encoding="utf-8") as stream:
         json.dump(content, stream, ensure_ascii=True)
     os.replace(path_swp, path)
+
+
+def _structure_path(directory: str, configuration):
+    structure_file = configuration.get("structure_file", None)
+    if not structure_file:
+        return None
+    return os.path.join(directory, "input", structure_file)
+
+
+def _conservation_type(configuration):
+    if "conservation" not in configuration:
+        return ConservationType.NONE
+    name = configuration["conservation"]
+    for conservationType in ConservationType:
+        if conservationType.value == name:
+            return conservationType
+    return ConservationType.NONE
+
+
+def _output_name(execution: Execution):
+    if execution.structure_file:
+        file = execution.structure_file
+        return file[file.rindex("/") + 1:file.rindex(".")]
+    elif execution.structure_code:
+        suffix = ("_" + ",".join(execution.chains)) if execution.chains else ""
+        return execution.structure_code + suffix
+    elif execution.structure_uniprot:
+        return execution.structure_uniprot
+    else:
+        return "prediction-without-name"
 
 
 if __name__ == "__main__":
